requirements.txt
urmacropkg
pynput
pydirectinput
pyinstaller
requests
winocr
